+++ packages/evm-connector/contracts/data-services/RapidDemoConsumerBase.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.4;
 import "../core/RedstoneConsumerNumericBase.sol";
 
 contract RapidDemoConsumerBase is RedstoneConsumerNumericBase {
-  function getUniqueSignersThreshold() public view virtual override returns (uint256) {
+  function getUniqueSignersThreshold() public view virtual override returns (uint8) {
     return 1;
   }
 
@@ -14,7 +14,7 @@ contract RapidDemoConsumerBase is RedstoneConsumerNumericBase {
     view
     virtual
     override
-    returns (uint256)
+    returns (uint8)
   {
     if (_signerAddress == 0xf786a909D559F5Dee2dc6706d8e5A81728a39aE9) {
       return 0;
